reform structure and folder names
