+++ policyengine_us/reforms/reforms.py
@@ -74,6 +74,10 @@ from .snap import (
 )
 from .states.dc.property_tax import create_dc_property_tax_credit_reform
 
+from .states.ny.inflation_rebates import (
+    create_ny_2025_inflation_rebates_reform,
+)
+
 
 from policyengine_core.reforms import Reform
 import warnings
@@ -166,6 +170,9 @@ def create_structural_reforms_from_parameters(parameters, period):
     dc_property_tax_credit = create_dc_property_tax_credit_reform(
         parameters, period
     )
+    ny_2025_inflation_rebates = create_ny_2025_inflation_rebates_reform(
+        parameters, period
+    )
 
     reforms = [
         afa_reform,
@@ -201,6 +208,7 @@ def create_structural_reforms_from_parameters(parameters, period):
         abolish_snap_deductions,
         abolish_snap_net_income_test,
         dc_property_tax_credit,
+        ny_2025_inflation_rebates,
     ]
     reforms = tuple(filter(lambda x: x is not None, reforms))
