+++ policyengine_us/reforms/states/ny/inflation_rebates/__init__.py
@@ -1,0 +1,3 @@
+from .ny_2025_inflation_rebates import (
+    create_ny_2025_inflation_rebates_reform,
+)
